+++ lib/adjust.dart
@@ -15,7 +15,7 @@ import 'package:adjust_sdk/adjust_event.dart';
 import 'package:adjust_sdk/adjust_attribution.dart';
 
 class Adjust {
-  static const String _sdkPrefix = 'flutter4.18.1';
+  static const String _sdkPrefix = 'flutter4.21.0';
   static const MethodChannel _channel = const MethodChannel('com.adjust.sdk/api');
 
   static void start(AdjustConfig config) {
@@ -53,6 +53,10 @@ class Adjust {
 
   static void gdprForgetMe() {
     _channel.invokeMethod('gdprForgetMe');
+  }
+
+  static void disableThirdPartySharing() {
+    _channel.invokeMethod('disableThirdPartySharing');
   }
 
   static void onResume() {